build(karma-runner-run.js): move into its own directory and make it a commonjs module
